+++ modules/canvaskit/CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    the encoding or serialization failed.
  - The `ShapedText` type has been removed. Clients who want ShapedText should use the
    Paragraph APIs.
+   
+### Removed
+ - `Font.measureText`, which had been previously deprecated. Clients should use either
+   Paragraph APIs or `Font.getGlyphWidths` instead (the latter does no shaping).
+ - `Font.getWidths`, which had been previously deprecated. Clients should use `Font.getGlyphWidths`.
 
 ### Type Changes (index.d.ts)
  - `Image.encodeToDataWithFormat` was incorrectly documented as its own thing.