+++ modules/canvaskit/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.24.0] - 2021-02-18
+
 ### Added
  - The Skottie factory (MakeManagedAnimation) now accepts an optional logger object.